+++ app/src/main/java/com/sefford/artdrian/common/di/ApplicationComponent.kt
@@ -2,9 +2,11 @@ package com.sefford.artdrian.common.di
 
 import com.sefford.artdrian.Artpplication
 import com.sefford.artdrian.connectivity.di.ConnectivityModule
+import com.sefford.artdrian.downloads.di.DownloadBoundariesModule
 import com.sefford.artdrian.downloads.di.DownloadsModule
 import com.sefford.artdrian.downloads.tasks.DownloadTask
 import com.sefford.artdrian.notifications.di.NotificationsModule
+import com.sefford.artdrian.wallpapers.di.WallpapersBoundaryModule
 import com.sefford.artdrian.wallpapers.di.WallpapersModule
 import com.sefford.artdrian.wallpapers.ui.detail.di.WallpaperDetailComponent
 import com.sefford.artdrian.wallpapers.ui.detail.di.WallpaperDetailModule
@@ -20,9 +22,11 @@ import javax.inject.Singleton
         ConcurrencyModule::class,
         ConnectivityModule::class,
         CoreModule::class,
+        DownloadBoundariesModule::class,
         DownloadsModule::class,
         NotificationsModule::class,
         NetworkModule::class,
+        WallpapersBoundaryModule::class,
         WallpapersModule::class,
     ]
 )
