+++ app/src/main/java/com/sefford/artdrian/common/di/CoreModule.kt
@@ -1,7 +1,5 @@
 package com.sefford.artdrian.common.di
 
-import com.sefford.artdrian.common.utils.DefaultLogger
-import com.sefford.artdrian.common.utils.Logger
 import dagger.Module
 import dagger.Provides
 import kotlinx.serialization.json.Json
@@ -16,11 +14,5 @@ class CoreModule {
         isLenient = true
         ignoreUnknownKeys = true
     }
-
-
-    @Provides
-    @Singleton
-    fun provideLogger(logger: DefaultLogger): Logger = logger
-
 }
