+++ app/src/test/java/com/sefford/artdrian/di/TestComponent.kt
@@ -1,14 +1,18 @@
 package com.sefford.artdrian.di
 
 import com.sefford.artdrian.common.di.CoreModule
+import com.sefford.artdrian.connectivity.di.ConnectivityModule
 import com.sefford.artdrian.downloads.data.datasources.DownloadsDataSourceTest
+import com.sefford.artdrian.downloads.di.DownloadsModule
 import com.sefford.artdrian.downloads.domain.model.DownloadProcessFetchTest
 import com.sefford.artdrian.downloads.domain.model.DownloadProcessPrimeTest
 import com.sefford.artdrian.downloads.domain.model.DownloadProcessProbeTest
 import com.sefford.artdrian.downloads.domain.model.DownloadProcessViabilityTest
+import com.sefford.artdrian.notifications.di.NotificationsModule
 import com.sefford.artdrian.wallpapers.data.datasources.WallpaperLocalDataSourceTest
 import com.sefford.artdrian.wallpapers.data.datasources.WallpaperNetworkDataSourceForListsTest
 import com.sefford.artdrian.wallpapers.data.datasources.WallpaperNetworkDataSourceForSingleWallpapersTest
+import com.sefford.artdrian.wallpapers.di.WallpapersModule
 import com.sefford.artdrian.wallpapers.ui.detail.di.WallpaperDetailComponent
 import com.sefford.artdrian.wallpapers.ui.detail.di.WallpaperDetailModule
 import com.sefford.artdrian.wallpapers.ui.list.WallpaperListViewModel
@@ -18,7 +22,12 @@ import javax.inject.Singleton
 @Component(
     modules = [
         CoreModule::class,
-        DoublesModule::class]
+        ConnectivityModule::class,
+        DownloadsModule::class,
+        DoublesModule::class,
+        NotificationsModule::class,
+        WallpapersModule::class,
+    ]
 )
 @Singleton
 interface TestComponent {
